+++ src/main/java/com/someguyssoftware/ddenizens/entity/monster/DenizensMonster.java
@@ -75,13 +75,22 @@ public abstract class DenizensMonster extends Monster implements IDenizensMonste
 				&& isDarkEnoughToSpawn(level, pos, random)
 				&& checkMobSpawnRules(mob, level, spawnType, pos, random);
 	}
-	
+
+	public static boolean checkDDMonsterCanSeeSkySpawnRules(EntityType<? extends Mob> mob, ServerLevelAccessor level, MobSpawnType spawnType, BlockPos pos, RandomSource random) {
+		Config.IMobConfig mobConfig = Config.Mobs.MOBS.get(EntityType.getKey(mob));
+		Config.CommonSpawnConfig config = mobConfig.getSpawnConfig();
+		return config.enabled.get()
+				&& level.getDifficulty() != Difficulty.PEACEFUL
+				&& isValidHeight(pos, config)
+				&& level.canSeeSky(pos)
+				&& checkMobSpawnRules(mob, level, spawnType, pos, random);
+	}
+
 	public static boolean isValidHeight(BlockPos pos, CommonSpawnConfig config) {
 		return pos.getY() > config.minHeight.get() && pos.getY() < config.maxHeight.get();
 	}
 	
 	public static boolean checkDDMonsterNetherSpawnRules(EntityType<? extends Mob> mob, ServerLevelAccessor level, MobSpawnType spawnType, BlockPos pos, RandomSource random) {
-//		DD.LOGGER.info("checking nether spawn rules at -> {}", pos);
 		IMobConfig mobConfig = Config.Mobs.MOBS.get(EntityType.getKey(mob));
 		NetherSpawnConfig config = ((INetherMobConfig)mobConfig).getNetherSpawn();
 		return config.enabled.get()