+++ src/main/java/com/someguyssoftware/ddenizens/setup/Registration.java
@@ -24,6 +24,9 @@ import com.google.common.collect.Lists;
 import com.someguyssoftware.ddenizens.DD;
 import com.someguyssoftware.ddenizens.capability.GhoulCapability;
 import com.someguyssoftware.ddenizens.entity.monster.*;
+import com.someguyssoftware.ddenizens.entity.monster.skeleton.FossilizedSkeleton;
+import com.someguyssoftware.ddenizens.entity.monster.skeleton.IronSkeleton;
+import com.someguyssoftware.ddenizens.entity.monster.skeleton.MagmaSkeleton;
 import com.someguyssoftware.ddenizens.entity.projectile.*;
 
 import com.someguyssoftware.ddenizens.item.*;
@@ -63,6 +66,13 @@ public class Registration {
 	public static final String BOULDER = "boulder";
 	public static final String ORC = "orc";
 	public static final String SKELETON_WARRIOR = "skeleton_warrior";
+	public static final String WINGED_SKELETON = "winged_skeleton";
+	public static final String FOSSILIZED_SKELETON = "fossilized_skeleton";
+	public static final String IRON_SKELETON = "iron_skeleton";
+	public static final String MAGMA_SKELETON = "magma_skeleton";
+	public static final String SKELETON_CHAMPION = "skeleton_champion";
+	public static final String DEATH_KNIGHT = "death_knight";
+
 	
 	// projectile names
 	private static final String PARALYSIS_SPELL = "slow";
@@ -71,10 +81,7 @@ public class Registration {
 	private static final String DISARM_SPELL = "disarm";
 	private static final String FIRESPOUT_SPELL = "firespout";
 	private static final String ROCK = "rock";
-	
-	// unused
-	public static Capability<GhoulCapability> GHOUL_CAPABILITY = null;
-	
+
 	/*
 	 * deferred registries
 	 */
@@ -96,14 +103,14 @@ public class Registration {
 			.build(BOULDER));
 	
 	public static final RegistryObject<EntityType<Headless>> HEADLESS_ENTITY_TYPE = Registration.ENTITIES.register(HEADLESS, () -> EntityType.Builder.of(Headless::new, MobCategory.MONSTER)
-			.sized(0.6F, 1.5F)
+			.sized(0.65F, 1.5F)
 			.clientTrackingRange(8)
 			.setShouldReceiveVelocityUpdates(false)
 			.setTrackingRange(50)
 			.build(HEADLESS));
 	
 	public static final RegistryObject<EntityType<Orc>> ORC_ENTITY_TYPE = Registration.ENTITIES.register(ORC, () -> EntityType.Builder.of(Orc::new, MobCategory.MONSTER)
-			.sized(0.6F, 1.95F)
+			.sized(1F, 1.99F)
 			.clientTrackingRange(12)
 			.setShouldReceiveVelocityUpdates(false)
 			.build(ORC));
@@ -170,6 +177,31 @@ public class Registration {
 			.clientTrackingRange(12)
 			.setShouldReceiveVelocityUpdates(false)
 			.build(SKELETON_WARRIOR));
+
+	public static final RegistryObject<EntityType<WingedSkeleton>> WINGED_SKELETON_TYPE = Registration.ENTITIES.register(WINGED_SKELETON, () -> EntityType.Builder.of(WingedSkeleton::new, MobCategory.MONSTER)
+			.sized(0.75F, 1.95F)
+			.clientTrackingRange(12)
+			.setShouldReceiveVelocityUpdates(false)
+			.build(WINGED_SKELETON));
+
+	public static final RegistryObject<EntityType<FossilizedSkeleton>> FOSSILIZED_SKELETON_TYPE = Registration.ENTITIES.register(FOSSILIZED_SKELETON, () -> EntityType.Builder.of(FossilizedSkeleton::new, MobCategory.MONSTER)
+			.sized(0.6F, 1.95F)
+			.clientTrackingRange(12)
+			.setShouldReceiveVelocityUpdates(false)
+			.build(FOSSILIZED_SKELETON));
+
+	public static final RegistryObject<EntityType<IronSkeleton>> IRON_SKELETON_TYPE = Registration.ENTITIES.register(IRON_SKELETON, () -> EntityType.Builder.of(IronSkeleton::new, MobCategory.MONSTER)
+			.sized(0.63F, 2.1F)
+			.clientTrackingRange(15)
+			.setShouldReceiveVelocityUpdates(false)
+			.build(IRON_SKELETON));
+
+	public static final RegistryObject<EntityType<MagmaSkeleton>> MAGMA_SKELETON_TYPE = Registration.ENTITIES.register(MAGMA_SKELETON, () -> EntityType.Builder.of(MagmaSkeleton::new, MobCategory.MONSTER)
+			.sized(0.63F, 2.1F)
+			.clientTrackingRange(15)
+			.setShouldReceiveVelocityUpdates(false)
+			.fireImmune()
+			.build(MAGMA_SKELETON));
 	
 	// projectile entities
 	public static final RegistryObject<EntityType<ParalysisSpell>> PARALYSIS_SPELL_ENTITY_TYPE =
@@ -235,7 +267,11 @@ public class Registration {
 	public static final RegistryObject<Item> DAEMON_EGG = Registration.ITEMS.register(DAEMON + "_egg", () -> new DaemonEggItem(DAEMON_ENTITY_TYPE, 0xff0000, 0xfc0000, new Item.Properties()));
 
 	public static final RegistryObject<Item> SKELETON_WARRIOR_EGG = Registration.ITEMS.register(SKELETON_WARRIOR + "_egg", () -> new SkeletonWarriorEggItem(SKELETON_WARRIOR_TYPE, 0xf5f6d2, 0xcdc3bb, new Item.Properties()));
+	public static final RegistryObject<Item> WINGED_SKELETON_EGG = Registration.ITEMS.register(WINGED_SKELETON + "_egg", () -> new WingedSkeletonEggItem(WINGED_SKELETON_TYPE, 0xf5f6d2, 0xcdc3bb, new Item.Properties()));
 
+	public static final RegistryObject<Item> FOSSILIZED_SKELETON_EGG = Registration.ITEMS.register(FOSSILIZED_SKELETON + "_egg", () -> new FossilizedSkeletonEggItem(FOSSILIZED_SKELETON_TYPE, 0xf5f6d2, 0xcdc3bb, new Item.Properties()));
+	public static final RegistryObject<Item> IRON_SKELETON_EGG = Registration.ITEMS.register(IRON_SKELETON + "_egg", () -> new IronSkeletonEggItem(IRON_SKELETON_TYPE, 0xf5f6d2, 0xcdc3bb, new Item.Properties()));
+	public static final RegistryObject<Item> MAGMA_SKELETON_EGG = Registration.ITEMS.register(MAGMA_SKELETON + "_egg", () -> new MagmaSkeletonEggItem(MAGMA_SKELETON_TYPE, 0x4b0000, 0xff7900, new Item.Properties()));
 
 	// projectiles
 	public static final RegistryObject<Item> PARALYSIS_SPELL_ITEM = Registration.ITEMS.register(PARALYSIS_SPELL, () -> new Item(new Item.Properties()));
@@ -287,6 +323,7 @@ public class Registration {
 	public static final RegistryObject<SoundEvent> AMBIENT_SHADOWLORD = registerSoundEvent("ambient_shadowlord");
 	public static final RegistryObject<SoundEvent> AMBIENT_SHADOW = registerSoundEvent("ambient_shadow");
 	public static final RegistryObject<SoundEvent> SHADOWLORD_STEP = registerSoundEvent("shadowlord_step");
+	public static final RegistryObject<SoundEvent> WINGED_SKELETON_FLAP = registerSoundEvent("winged_skeleton_flap");
 
 	// NOTE must add mob to ALL_MOBS collection in order to register them to the biomes - see CommonSetup.onBiomeLoading
 	static {
@@ -302,6 +339,8 @@ public class Registration {
 		ALL_MOBS.add(DAEMON_ENTITY_TYPE);
 		ALL_MOBS.add(ORC_ENTITY_TYPE);
 		ALL_MOBS.add(SKELETON_WARRIOR_TYPE);
+		ALL_MOBS.add(WINGED_SKELETON_TYPE);
+		ALL_MOBS.add(FOSSILIZED_SKELETON_TYPE);
 	}
 	
 	/**