+++ CHANGELOG.md
@@ -6,7 +6,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.0] - 2024-02 - Shadowlord Update
+## [1.4.0] - 2024-02 - Skeleton Update
+
+### Changed
+
+- Ghoul had RestrictSunGoal twice - removed duplicate.
+- Increased the hit box on the Orc to match body size better.
+- Updates to DD Bestiary.
+
+### Added
+
+- Winged Skeleton.
+- Fossilized Skeleton.
+- Iron Skeleton.
+- Magma Skeleton.
+- PassiveMeleeAttackGoal.
+- VariantPoweredRangedBowAttackGoal.
+
+
+## [1.3.0] - 2024-01-31 - Shadowlord Update
 
 ### Changed